backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-genai>=1.0.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.160.0
pandas>=2.2.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pydantic>=2.10.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
PyJWT>=2.8.0
